\name{MCEMCRJointCPP-package}
\alias{MCEMCRJointCPP-package}
\alias{MCEMCRJointCPP}
\docType{package}
\title{
\packageTitle{MCEMCRJointCPP}
}
\description{
\packageDescription{MCEMCRJointCPP}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{MCEMCRJointCPP}

Maintainer: \packageMaintainer{MCEMCRJointCPP}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = MCEMCRJointCPP)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{MCEMCRJointCPP}}
%\section{Documentation Index}{\packageIndices{MCEMCRJointCPP}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
